feat(compiler): updated vm instructions and design

+++ compiler/fabc_vm/src/value.rs
@@ -3,7 +3,7 @@ use std::rc::Rc;
 #[derive(Clone)]
 pub enum Value {
     String(Rc<String>),
-    Number(f64),
+    Number(i64),
     Boolean(bool),
     None,
 }
